Initial server configuration file in Go
This commit creates a new Go file called main.go that starts a local
server at port number configured by $PORT. $PORT is needed as web dynos
on Heroku set the $PORT variable.

Handler functions have been created in main.go that allow for
HTTP GET and POST methods to be handled. Each one writes the the
http.ResponseWriter, which is then sent to the client.

A Procfile has also been created to tell Heroku what command to run to
start the server. 'web' is needed to tell Heroku that the application is
a web application.