Add new functions to check for aggregation of radius
A counter is used to keep track of how many notes we have inserted into
the notes table during the run time of the server.

The handler checks this counter, and for every 3rd insert, it will
attempt to perform an aggregation with the new note.

The first step of the aggregation is to get all notes that surround the
new note within a certain radius. Here we have chosen for 50m. This
requires calculating how far each note is by their longitude and
latitude, which requires conversion to the spherical distance of two
points on a sphere.