Add function to check if two notes are within a time frame
New function withinTimeFrame will check if two dates are within a
time frame of each other. In order for this to work, we need to parse
the time strings from the note struct, and we need to make sure that the
time zones of the two dates are the same. This is ensured by changing the
SQL query of getAllNotes to specify that the time zone of startTime and
endTime is to be returned in UTC.

Also:
- began writing the pipeline function of the note aggregator for
getting notes that are within 15 minutes of the start and end times of
the new note.
- Go plugin fixed a load of tabbing