Add test for ConstructAggregatedNote
Test has been added to db_test.go to test the implementation of
ConstructAggregatedNote function in models/note.go with the current
policies. As part of this, a few private helper methods have been
created to simplify creating a user or checking if an id is contained in
an array of ids. A constant TOLERANCE has also been added for the
purpose of checking if floating point numbers are equal or not. This is
used for checking the return values of the latitude and longitude of the
new aggregated note.

The return values of the aggregate helper functions have been changed to
addresses as the fields of a note struct are all pointers.