Connect the whole pipeline for note aggregation
handler/note.go now calls the ConstructAggregatedNote function. There
used to be an if check inside the function to check if the notes array
passed in was empty, however this has been moved to the handler to check
instead. The reason for this is because the handler has the context and
so can return right away if the notes array is empty, as opposed to needing
to check for a nil return value.

Also:
- changed the return type of mergeNotes to be (int64, error), because
  it calls createNote which has that return type.
- Set the Id field of the note struct to be int64 to be consistent with
  what we've been using for ids in other functions.
- Added the newId that was returned in the first createNote call to the
  note, as we will need this later on to remove it again.